backend/requirements.txt
fastapi[standard]
uvicorn
python-multipart
langchain
langchain-openai
pypdf
python-docx
pydantic
python-dotenv

